Move #distance_to to User, delete Distanceable (because not needed for Dog)
